evaluateLevel from Float to BigDecimal

+++ tm-steer-api/src/main/groovy/cn/edu/bnuz/steer/ObservationFormCommand.groovy
@@ -19,7 +19,7 @@ class ObservationFormCommand {
     Integer                 attendantStds
     Integer                 lateStds
     Integer                 leaveStds
-    Float                   evaluateLevel
+    BigDecimal              evaluateLevel
     List<EvaluationsItem>   evaluations
     String                  evaluationText
     String                  suggest
